+++ tests/benchmarking/test_run_benchmark.py
@@ -1,37 +1,5 @@
-from itertools import product
-
-import numpy as np
-import pytest
-from estimagic.benchmarking.benchmarking import get_benchmark_problems
-from estimagic.benchmarking.benchmarking import run_benchmark
-
-PARMETRIZATION = []
-for name in ["more_wild", "cartis_roberts"]:
-    for additive, multiplicative in product([False, True], repeat=2):
-        PARMETRIZATION.append((name, additive, multiplicative))
-
-
-@pytest.mark.parametrize("name, additive_noise, multiplicative_noise", PARMETRIZATION)
-def test_get_problems(name, additive_noise, multiplicative_noise):
-    is_noisy = any((additive_noise, multiplicative_noise))
-    problems = get_benchmark_problems(
-        name=name,
-        additive_noise=additive_noise,
-        multiplicative_noise=multiplicative_noise,
-    )
-    first_name = list(problems)[0]
-    first = problems[first_name]
-    func = first["inputs"]["criterion"]
-    params = first["inputs"]["params"]
-
-    np.random.seed()
-    first_eval = func(params)["value"]
-    second_eval = func(params)["value"]
-
-    if is_noisy:
-        assert first_eval != second_eval
-    else:
-        assert first_eval == second_eval
+from estimagic import get_benchmark_problems
+from estimagic.benchmarking.run_benchmark import run_benchmark
 
 
 def test_run_benchmark_dict_options(tmpdir):
